refactor: consolidate tests into single test/ directory
- Remove old vows test files from test/ directory (preserve data dirs)
- Move mocha tests from mocha-tests/ to test/ directory
- Remove empty mocha-tests/ directory
- Update test file paths to work from new location
- Update package.json test script to point to test/**/*.js
